Redid all data files

+++ src/PocketData/Trait.cs
@@ -4,9 +4,13 @@ namespace PocketData;
 
 public class Trait
 {
-	[JsonPropertyName("code")] public int Code { get; set; }
+	[JsonPropertyName("code")]
+	public int Code { get; set; }
 
-	[JsonPropertyName("dataId")] public int DataId { get; set; }
+	[JsonPropertyName("dataId")]
+	public int DataId { get; set; }
 
-	[JsonPropertyName("value")] public int Value { get; set; }
+	[JsonPropertyName("value")]
+	public double Value { get; set; }
+
 }
